Add files via upload
This is a java GUI program that about MouseListener. It contains a GUI and each of the mouseListener method has their own labels, and then when it clicks, release, exited, pressed and entered, it will change the background color of the label  and prints how many time the action happens.